fix: project structure and comments on render functions
