--- app/layout/containers/entry.ejs
@@ -1,47 +1,0 @@
-<div class="container mt-4">
-  <div class="jumbotron mb-5 p-3">
-    <h1 class="text-center mb-3 font-weight-bold pl-0">Try onsite-demo</h1>
-    <hr class="my-4">
-    <form class="form-horizontal" method="post">
-      <div class="row justify-content-md-center">
-        <div class="form-group m-2">
-          <label for="apiKey" class="text-left col-md-4 font-weight-bold pl-0">
-            ApiKey<span class="text-danger">*</span>
-          </label>
-          <input class="input-md textinput form-control" id="apiKey" type="text" />
-        </div>
-        <div class="form-group m-2">
-          <label for="secretKey" class="text-left col-md-4 font-weight-bold pl-0">
-            SecretKey<span class="text-danger">*</span>
-          </label>
-          <input class="input-md textinput form-control" id="secretKey" type="text" />
-        </div>
-        <div class="form-group m-2">
-          <label for="name" class="text-left col-md-4 font-weight-bold pl-0">
-            Name<span class="text-danger">*</span>
-          </label>
-          <input class="input-md textinput form-control" id="name" type="text" />
-        </div>
-      </div>
-      <div class="row justify-content-md-center">
-        <div class="form-group m-2">
-          <label for="source" class="text-left col-md-4 font-weight-bold pl-0">
-            Source<span class="text-danger">*</span>
-          </label>
-          <input class="input-md textinput form-control" id="source" type="text" />
-        </div>
-        <div class="form-group m-2">
-          <label for="deviceId" class="text-left col-md-4 font-weight-bold pl-0">
-            DeviceId<span class="text-danger">*</span>
-          </label>
-          <input class="input-md textinput form-control" id="deviceId" type="text" />
-        </div>
-      </div>
-      <div class="row justify-content-md-center mt-2">
-        <div class="col-md-auto">
-          <button id="try_btn" type="button" class="btn btn-primary">Try it</button>
-        </div>
-      </div>
-    </form>
-  </div>
-</div>
