--- app/layout/containers/json.ejs
@@ -1,5 +1,0 @@
-<hr class="my-4 mt-3">
-<h1><b>Json file:</b></h1>
-<div class="row p-1">
-  <pre id="container-json" class="json-file flex-row flex-column m-4 p-1"></pre>
-</div>
