--- app/layout/containers/widgets.ejs
@@ -1,9 +1,0 @@
-<hr class="my-4 mt-3">
-<h1>
-  <b>Divs de vitrines:</b>
-</h1>
-<div class="container mb-4 mt-4">
-  <div class="top mt-2 border pl-5 pr-5 pt-3 pb-5 rounded"></div>
-  <div class="middle mt-2 border pl-5 pr-5 pt-3 pb-5 rounded"></div>
-  <div class="bottom mt-2 border pl-5 pr-5 pt-3 pb-5 rounded"></div>
-</div>
