--- app/layout/index.ejs
@@ -1,14 +1,0 @@
-<html>
-  <head>
-    <% include containers/header %>
-  </head>
-
-  <body>
-    <% include containers/entry %>
-    <div id="container-demo" class="d-none container mt-2">
-      <% include containers/widgets %>
-      <% include containers/json %>
-    </div>
-    <% include containers/footer %>
-  </body>
-</html>
