--- app/layout/templates/product.ejs
@@ -1,16 +1,0 @@
-
-<%
-  const recs = widget.displays[0].recommendations;
-  let product;
-
-  Object.keys(recs).forEach((index) => {
-    product = recs[index];
-%>
-    <div id="<%= product.id %>" class="row justify-content-center product-slide pt-4">
-      <img src="https://<%= product.images.default %>" class="rounded mx-auto d-block">
-      <div class="m-1"><%= product.name %></div>
-      <div class="m-1 font-weight-bold"><%= product.price %></div>
-    </div>
-<%
-  });
-%>
