+++ layout/index.html
@@ -1,0 +1,86 @@
+<html>
+  <head>
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous">
+    <link rel="stylesheet" type="text/css" href="//cdn.jsdelivr.net/gh/kenwheeler/slick@1.8.1/slick/slick.css"/>
+    <link rel="stylesheet" type="text/css" href="//cdn.jsdelivr.net/gh/kenwheeler/slick@1.8.1/slick/slick-theme.css"/>
+    <link href="https://fonts.googleapis.com/css?family=Roboto" rel="stylesheet">
+    <link rel="stylesheet" type="text/css" href="https://cdnjs.cloudflare.com/ajax/libs/bootswatch/4.1.3/pulse/bootstrap.css"/>
+  </head>
+
+  <body>
+    <div class="container mt-4">
+      <div class="jumbotron mb-5 p-3">
+        <h1 class="text-center mb-3 font-weight-bold pl-0">Try onsite-demo</h1>
+        <hr class="my-4">
+        <form class="form-horizontal">
+          <div class="row justify-content-md-center">
+            <div class="form-group m-2 col-md-3">
+              <label for="apiKey" class="text-left font-weight-bold pl-0">
+                ApiKey<span class="text-danger">*</span>
+              </label>
+              <input class="input-md textinput form-control" id="apiKey" type="text" />
+            </div>
+            <div class="form-group m-2 col-md-3">
+              <label for="secretKey" class="text-left font-weight-bold pl-0">
+                SecretKey<span class="text-danger">*</span>
+              </label>
+              <input class="input-md textinput form-control" id="secretKey" type="text" />
+            </div>
+            <div class="form-group m-2 col-md-3">
+              <label for="name" class="text-left font-weight-bold pl-0">
+                Name<span class="text-danger">*</span>
+              </label>
+              <input class="input-md textinput form-control" id="name" type="text" />
+            </div>
+          </div>
+          <div class="row justify-content-md-center">
+            <div class="form-group m-2 col-md-3">
+              <label for="source" class="text-left font-weight-bold pl-0">
+                Source<span class="text-danger">*</span>
+              </label>
+              <input class="input-md textinput form-control" id="source" type="text" />
+            </div>
+            <div class="form-group m-2 col-md-3">
+              <label for="deviceId" class="text-left  font-weight-bold pl-0">
+                DeviceId<span class="text-danger">*</span>
+              </label>
+              <input class="input-md textinput form-control" id="deviceId" type="text" />
+            </div>
+          </div>
+          <div class="row justify-content-md-center mt-2">
+            <button id="try_btn" type="button" class="btn btn-primary">Try it</button>
+          </div>
+        </form>
+      </div>
+    </div>    
+
+    <div id="container-demo" class="d-none container mt-2">
+      <hr class="my-4 mt-3">
+      <h1>Slots de vitrines:</h1>
+      <div class="container mb-4 mt-4">
+        <div class="mt-2 border p-3 rounded">
+          <h3 class="font-weight-bold">Top</h3>
+          <hr class="my-4">
+          <div class="top"></div>
+        </div>
+        <div class="mt-2 border p-3 rounded">
+          <h3 class="font-weight-bold">Middle</h3>
+          <hr class="my-4">
+          <div class="middle"></div>
+        </div>
+        <div class="mt-2 border p-3 rounded">
+          <h3 class="font-weight-bold">Bottom</h3>
+          <hr class="my-4">
+          <div class="bottom"></div>
+        </div>
+      </div>
+
+      <hr class="my-4 mt-3">
+      <h1 class="font-weight-bold">Json file:</h1>
+      <pre id="container-json" class="json-file flex-row flex-column m-4 p-1"></pre>
+    </div>
+
+    <script src="https://code.jquery.com/jquery-1.12.4.min.js" integrity="sha256-ZosEbRLbNQzLpnKIkEdrPv7lOy9C27hHQ+Xp8a4MxAQ=" crossorigin="anonymous"></script>
+    <script type="text/javascript" src="//cdn.jsdelivr.net/gh/kenwheeler/slick@1.8.1/slick/slick.min.js"></script>
+  </body>
+</html>
