+++ src/styles/style.scss
@@ -33,12 +33,6 @@
   box-shadow: 0px 0px 55px -10px rgba(0,0,0,0.14);
 }
 
-pre {
-  outline: 1px solid #ccc; 
-  padding: 5px; 
-  margin: 5px; 
-}
-
 .string { 
   color: green; 
 }
@@ -62,4 +56,7 @@ pre {
 .json-file {
   height: 800px;
   background-color: #f9f9f8;
+  outline: 1px solid #ccc; 
+  padding: 5px; 
+  margin: 5px; 
 }
